backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-bcrypt==0.7.1
flask-jwt-extended==4.3.1
numpy==1.22.0
textblob==0.15.3
joblib==1.1.0
google-generativeai==0.3.1
